agents/forgery_evidence_checker/requirements.txt
kaggle 
transformers
torch 
scikit-learn 
accelerate
google-generativeai
google.adk